mm: kmemleak: Don't die when memory allocation fails
When memory is leaking, it's going to be harder to allocate more memory,
making it more likely for this failure condition inside of kmemleak to
manifest itself. This is extremely frustrating since kmemleak kills
itself upon the first instance of memory allocation failure.

Bypass that and make kmemleak more resilient when memory is running low.

Change-Id: I5ab49726e3bae76ece0cfcccb7b63123d10b5782
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: techyminati <sinha.aryan03@gmail.com>
Signed-off-by: Momo Belia Deviluke <fajarslebew31@gmail.com>
Signed-off-by: Riko Dan Pajar <christiantoriko227@gmail.com>